requirements.txt
aiogram==3.7.0
asyncpg==0.29.0
python-dotenv==1.0.1
aiohttp==3.9.5
google-generativeai==0.7.1
apscheduler==3.10.4
SQLAlchemy==2.0.30
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml[html_clean]==5.2.2
pytz==2024.1